Removes ServerInterface and moves its functionality elsewhere
Up until now, the plugin used the ServerInterface to manage getting the server version and managing the whitelist. In an effort to make it so that the plugin will eventually move off NMS and onto the standard Spigot/Paper API, I've removed the ServerInterface completely and moved its functions onto something that doesn't require NMS to get server versions or manage the whitelist.